refactor: more folder sorting, move legacy css
