--- src/SocialMediaFooter.css
@@ -1,21 +1,0 @@
-.social-media-footer {
-    background: rgb(86, 190, 167);
-    width: 100vw;
-    height: 10vh;
-    position: fixed;
-    bottom: 0;
-}
-
-.social-media-list {
-    text-align: center;
-    margin: 0;
-}
-
-.social-media-link {
-    display: inline-block;
-    padding: 20px;
-}
-
-.social-media-icon {
-    width: 30px;
-}
