--- src/chart.css
@@ -1,9 +1,0 @@
-
-.arc text {
-    font: 10px sans-serif;
-    text-anchor: middle;
-  }
-
-  .arc path {
-    stroke: #fff;
-  }
