+++ src/components/App.js
@@ -1,0 +1,16 @@
+import React, { Component } from "react";
+import Nav from "./Nav";
+import Footer from "./Footer";
+
+class App extends Component {
+  render() {
+    return (
+      <div className="content-wrapper">
+        <Nav />
+        <Footer />
+      </div>
+    );
+  }
+}
+
+export default App;
