+++ src/index.css
@@ -1,3 +1,5 @@
+/* Moving all initial CSS as this will probably all change so let's keep the legacy stuff in one place */
+
 body {
   margin: 0;
   height: 100%;
@@ -9,7 +11,87 @@ body {
   -moz-osx-font-smoothing: grayscale;
 }
 
+p {
+  margin: 0;
+}
+
 code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
     monospace;
+}
+
+/* navigation */
+
+.main-content {
+  background: rgb(107, 229, 185);
+  height: 100vh;
+  width: 100vw;
+  text-align: center;
+}
+
+.nav-container {
+  background: rgb(79, 126, 229);
+  display: flex;
+  justify-content: space-between;
+  padding: 2vmin;
+}
+
+.navigation-link-container {
+}
+
+.nav-link {
+  color: black;
+}
+
+.nav-link:active,
+.nav-link:visited {
+  color: lightblue;
+}
+
+/* chart */
+
+.arc text {
+  font: 10px sans-serif;
+  text-anchor: middle;
+}
+
+.arc path {
+  stroke: #fff;
+}
+
+/* Social Media Footer */
+
+.social-media-footer {
+  background: rgb(86, 190, 167);
+  width: 100vw;
+  height: 10vh;
+  position: fixed;
+  bottom: 0;
+}
+
+.social-media-list {
+  text-align: center;
+  margin: 0;
+}
+
+.social-media-link {
+  display: inline-block;
+  padding: 20px;
+}
+
+.social-media-icon {
+  width: 30px;
+}
+
+/* Events */
+
+.calendar-wrapper {
+  width: 400px;
+  float: right;
+}
+
+/* Team */
+
+.team-member-photo {
+  width: 300px;
 }
