--- src/pages/Events.css
@@ -1,4 +1,0 @@
-.calendar-wrapper {
-    width: 400px;
-    float: right;
-}
