--- src/pages/Team.css
@@ -1,3 +1,0 @@
-.team-member-photo {
-    width: 300px;
-}
